+++ changelog.md
@@ -14,6 +14,12 @@
   * Disabled "Prevent boating and minecarting"
 * Added Slimes module and feature
   * Slimes can now spawn bigger and jump more often
+* Digger Zombies -> Miner Mobs
+  * Miner's AI can now be applied to any mob in an entity eype tag (`enhancedai:can_be_miner`) (by default, only zombies, like before)
+  * Blocks can now be blacklisted with a block tag (`enhancedai:miner_block_blacklist`)
+  * Added 'Always require proper tool'
+  * Added `enhancedai:time_to_break_multiplier` nbt tag that can be used to set the multiplier on time to break blocks
+  * Upgraded the Wooden Pickaxe to a Stone Pickaxe
 * Mobs are now affected by blindness (They see only 10% of range)
 * Drowned should get stuck less
 * Witches no longer heal themselves if above 70% health